+++ CHANGELOG.md
@@ -7,6 +7,10 @@ _released 00 September 2025_
 - Features:
   - Assignments that have titles that start or end with newlines will get those
     trimmed. No more needless empty lines! (looking at you, $GUSH_TEACHER)
+- Improvements:
+  - Graded assignments will have a green outline around them. The default
+    background color for graded assignments is reverted to be the same as
+    "completed".
 
 ## v0.7.4